+++ docs/changelog.rst
@@ -39,6 +39,9 @@ Other little fixes:
 * Fix a potential crash when using image resizing with the
   :doc:`/plugins/fetchart` or :doc:`/plugins/embedart` without ImageMagick
   installed.
+* :doc:`/plugins/fetchart`: The ``beet fetchart`` command can now associate
+  local images with albums (unless ``--force`` is provided). Thanks to
+  brilnius.
 
 
 1.3.2 (December 22, 2013)